backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.0.0
geopy>=2.4.0
requests>=2.32.0
python-dotenv>=1.0.0
polyline>=2.0.0


